Initial commit of the "MultiTextInput" project.
MultiTextInput is a component which mimics functionality seen in Mail.app's "To:" filed, or Facebook's <fb:multi-friend-input> widget.
Essentially the component allows the user to type some text, hit enter, and have that text be 'added' to the component.  Users may then
remove existing text by pressing a visual x button, or hitting the backspace button.


git-svn-id: file:///home/igor/dev/stuff/svnbackup/trunk/wicketstuff-core@4681 ef7698a4-5110-0410-9fc6-c7eb3693863f