datatable-autocomplete related changes:
Created PrefixTrieMatch and AnyWhereTrieMatch to encapsulate the TrieNode<C> or Set<TrieNode<C>> that match an initial search.
The plan is to add in a Match Cache so that subsequent searches can simply search within the existing match versus starting from scratch each time.

The TrieMatch classes internalize the generation of the List<C> getWordList(int limit)

Added the ITrieDataProviderHints interface and default implementation to allow setting an upper limit on the number of results to fetch when the search string is short. (i.e. instead of building 20,000 element list when they type 'a' limit to something smaller like 500)

git-svn-id: file:///home/igor/dev/stuff/svnbackup/trunk/wicketstuff-core@5238 ef7698a4-5110-0410-9fc6-c7eb3693863f