wicketstuff-core related:
Seperated the new profile for oss.sonatype.org snapshots into two.

Now there is one for hudson to use to release to the repo
and an always on version to lookup snapshots from the sonatype repo.

The activator may need to change when we do a real release of 
wicketstuff-core on the sonatype infrastructure.



git-svn-id: file:///home/igor/dev/stuff/svnbackup/trunk/wicketstuff-core@5258 ef7698a4-5110-0410-9fc6-c7eb3693863f