datatable-autocomplete-related:
Building a suffix Trie within a prefix tree is very space inefficient.  

Committing the start of a TernarySearchTrie that is supposed to require a lot less nodes (and memory).

Adding in JUNG dependencies to the test scope of datatable_autocomplete to help visualize how the trie's are being constructed and to make sure they resemble trees more so than lists.

git-svn-id: file:///home/igor/dev/stuff/svnbackup/trunk@5308 ef7698a4-5110-0410-9fc6-c7eb3693863f