fixes the org.eclipse.jetty dependency in push-parent/push-examples that was breaking the build.
The  ${jetty.version} property redefined in the push-examples pom.xml clashed with the parent jetty plugin dependency since there are no 
org.mortbay.jetty artifacts for the 7.x.y series.

The fix is to rename the property to org.eclipse.jetty and update the dependency versions.  This approach was already taken with the 
wicket-html5-parent/pom.xml so the second part of the fix is to push up the definition of the property into the wicketstuff-core/pom.xml
  


git-svn-id: file:///home/igor/dev/stuff/svnbackup/trunk/wicketstuff-core@5439 ef7698a4-5110-0410-9fc6-c7eb3693863f