datatable-autocomplete changes: support onError
When using an AjaxButton for example and the form does not submit due to
validation errors there is an onError(target, form) that needs to be
implemented.

These changes add in a new method on the renamed IFormSubmitAction to support
handling the onError case.  It is connected to the DTAButton.onError()
callback.

git-svn-id: file:///home/igor/dev/stuff/svnbackup/branches/wicketstuff-core-1.4@5683 ef7698a4-5110-0410-9fc6-c7eb3693863f