deleted more files denoted by new .gitignore
I did 'cat /.gitignore | while read F; do find . -name $F; done;' and found that there were some other files
that should also have been deleted.

This commit contains the rest of the files that should be removed.

This excludes some request target classes from the minis project that are not related to the build /target directory.

It also reverts some of the changes to the profiles in the pom.xml that broke the build.