change wicket dependency to wicket-core
The prevously reverted change was correct but I failed to adjust the project pom.xml's that directly refer to it.

This commit changes the dependency from wicket (which no longer exists) to wicket-core.  And it also changes the project pom's that
used to refer to <artifactId>wicket</artifactId> to nothing since it is inherited from the parent project.