remove resources sections from project poms
The resources section can be inherited down from the root pom.

Also many of the plugins can also be inherited down.

Things like the maven-jetty-plugin and maven-eclipse-plugin don't need to be defined at all levels.

This commit removes all of the <resources> and <testresources> from the project poms
and lets them inherit the root pom definition.

I've left the scala pom alone because it does have a variation (including the scala files).