datatable-autocomplete: fix IE8 encoding issue with AJAXAutoCompleteBehavior
IE8 does not properly encode accented characters (i.e. with umlauts) when using
a GET request.  This commit wraps the auto complete field value with the
encodeURI javascript function to solve the issue.

Without this patch jetty 7.4.x dies when reading the request and before it is
handed off to wicket.