[datatable-autocomplete] Enable custom CSS resources for the overlay and table
The CSS classes on the rendered markup remains the same but now we support
custom CSS resources being provided.

This was added because in my work application the default style against the
root elements impacted the overlay panel.

With this fix I can write a custom stylesheet to cancel out the impact of the
main style.