[jdk-1.5-parent] crosscompile modules, java 5 compile for tests
The trick is that the java -source 1.5 and -target 1.5 are required for the maven plugin but the java 6 compiler lies
and still supports java 6 constructs in the source code.

So for safety we still build the tests using java 5 and run them using java 5 but for the main module code we cross compile
from java 6 to java 5.

To get the java 5 compiler to be used for the tests it needs to be forked.

I tested this by putting this class into both the main/java and test/java directories of one of the modules:

public class Simple {

        public static interface T {
                public void test();
        };

        private class TT implements T {

                @Override
                public void test() {
                }
        }

}

This compiles correctly if the compiler is java 6 but will fail if the compiler is java 5.