[spring-webflow-parent] Fix pom to inhereit from parent
This defines the parent section so that the spring-webflow-parent can inherit
the core/pom.xml and jdk-1.5-parent/pom.xml dependency and build settings.

I also remove dependencyManagement entries that are defined in the parent pom's.

And the compiler plugin settings since that is defined in the jdk-1.5-parent pom.