[openlayers] Add Java abstraction API for SelectFeature control.
The SelectFeature control is used to connect Javascript functionality to the
selection of features in a vector layer.

This commit adds in a new interface and abstract implementation that allow the
functionality to specified nicely in java and have the Control class itself
take care of the actual javascript constructor wiring.

For compatibility reasons the bare constructor remains available