[wicket-security-parent/swarm-parent] Fix test case from issue #84
I am not a user of the swarm or wicket-security project so it could be that this change is incorrect.  However when running the test in eclipse it looks like the page transition request is denied properly.  However while control used to go to the AccessDeniedPage it now stays put on the current page.

I have modified the GeneralTest.testMultiLogin() method to just check that the original page is stayed on versus showing the access denied page.