Introduce a separate sort property type parameter
We cannot assume that the sort type will be the same as the cell type. The data provider is responsible for sorting according to the one sort type it can sort with. If the sort type is the same as the cell type, then we cannot mix columns with different cell types using the same data provider.