Add support for servlet 3.0 security integrated with wicket-auth-roles
This commit adds support for using servlet 3.0 authenticate and
authorization that is configured using the servlet security container.
The commit also incudes a general purpose auto mount library that
generates code at compile time for pages marked with an annotation.
Since a runtime feature of this paradigm existed in WicketStuff, that
code is also merged into the new wicket-mount library.  The servlet 3.0
Security also uses the core features of the wicket-mount to enable
developers a solution to generate mounted pages that the servlet
container will intercept so that the configured authentication via
web.xml  is enforced.  Examples are included for testing the features.