Added transparent transformation to points.
All points that do not originate from a EPS:900913 projection need to be
transformed for an EPS:900913 projection before being placed on a
map. Specifically this effects OSM, Google and Bing points: these all
originate from an EPS:4326 (EPS:3857) or a "spherical mercator"
projection. The new functionality allows for the creation of LonLat
instances with a native and target projection, they are transformed when
rendered to Javascript.

A new SphericalMercatorLonLat class has also been created to make this
easier. All of these instances already have their native projection set
to EPS:4326 and their target projection set to EPS:900913.