Increased relevant statistics in admin page on reset.
There was a problem with a commonly added monitor which was never hit.
In that case all collected statistics have been in an initial state.
There have been some unexpected display issues with initial state.

But imho there is no need to clutter statistics with monitors never
recorded anything.

All was required in reset workflow to.