[offline-mode] Minor improvements
Store the cacheable entries in the application metadata.
Use eager singletons instead of synchronizing (on the class!) for each #getInstance() call.
Add namespace to the mount paths. Used '-' instead of '/' because otherwise the url resolving breaks.
Use ResourceReference#getDependencies() to make sure that ServiceWorker is always used with ServiceWorkerRegistration.
Rename OfflineCache#load() to #renderHead() to be in sync with all other similar methods.