Fix the delete action when there are two editable grids on one page.
When a page holds two editable grids, a delete event on the second grid will be performed by the first grid. First and second positions are in the order in which they were added to the page. This happens because the delete action is propagated as a message, but this message does not hold the object that receives the event. Thus the first EditableDataTable that receives the message will process it. I have added the receiving EditableDataTable objevt into the message payload and added a check for it  in the event processing.

- GridOperationData.java: Added a field to the GridOperationData payload to hold the EditableDataTable that receives the onClick event
- EditableDataTable.java: Added a check in the onEvent method in EditableDataTable so that each object will only process its own events.
- EditableGridActionsPanel: Added the receiving datatable object to the payload in EditableGridActionsPanel