Added component and behavior sub-classes to provide lambda handlers.
In cases where there is a single or dominant handler (e.g., onClick),
the handler can be specified in both a constructor and also in a fluent
setter (e.g., returning this). 

In cases where there are multiple meaningful handlers (e.g., onSubmit
and onError), the handlers must be specified using fluent setters. 

The normally overridden onXXX methods are marked as final and
implemented by calling the lambda handler.

Failure to specify a dominant handler will raise a WicketException. In
general, not specifying an _error_ handler does _not_ raise an error.